requirements.txt
robotframework==4.1.2
robotframework-seleniumlibrary==5.1.3
requests==2.25.0
